requirements.txt


python-dotenv
requests
mimetypes; python_version < "3.12"  # (builtin in 3.12+, so safe to skip if on 3.12)
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
langgraph>=0.0.25
faiss-cpu
openai
unstructured
unstructured[excel]
pdfminer.six

# backend/requirements.txt - Backend Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis for task queue and caching
redis[asyncio]==5.0.1
aioredis==2.0.1

# Background tasks and async support
celery==5.3.4
asyncio-mqtt==0.16.1

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Monitoring and metrics
psutil==5.9.6
prometheus-client==0.19.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
pydantic==2.5.2
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# WebSocket support
websockets==12.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Additional utilities
loguru==0.7.2
click==8.1.7
typer==0.9.0